selectMenuView titleView changed
#8
selectMenuView의 titleView를 collectionView의 header로 포함시킴
